rename : 검색 페이지 폴더 이동
